Social Media Prepper/requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
pillow>=10.0.0
python-dateutil>=2.8.0
